refactor: Refer to the correct models
