+++ nisystemlink/clients/file/models/__init__.py
@@ -1,4 +1,3 @@
-# from ._file_models import *
 from ._uploaded_file_info import UploadedFileInfo
 from ._delete_multiple import DeleteMutipleRequest
 from ._file_query_response import FileQueryResponse
